Refactor the ipam controllers into a seperate package per review. Also refactor all controller tests into common package
