+++ test/functional/clusteripamclaim_controller_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package controller
+package functional
 
 import (
 	. "github.com/onsi/ginkgo/v2"
@@ -24,6 +24,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	kcm "github.com/K0rdent/kcm/api/v1alpha1"
+	"github.com/K0rdent/kcm/internal/controller/ipam"
 )
 
 var _ = Describe("ClusterIPAMClaim Controller", func() {
@@ -39,7 +40,7 @@ var _ = Describe("ClusterIPAMClaim Controller", func() {
 		Expect(k8sClient.Update(ctx, resource)).To(Succeed())
 
 		By("Reconciling the resource after update")
-		reconciler := &ClusterIPAMClaimReconciler{Client: k8sClient, Scheme: k8sClient.Scheme()}
+		reconciler := &ipam.ClusterIPAMClaimReconciler{Client: k8sClient, Scheme: k8sClient.Scheme()}
 		_, err := reconciler.Reconcile(ctx, reconcile.Request{NamespacedName: types.NamespacedName{Name: name, Namespace: namespace}})
 		Expect(err).NotTo(HaveOccurred())
 
@@ -73,7 +74,7 @@ var _ = Describe("ClusterIPAMClaim Controller", func() {
 
 	baseReconciliation := func(namespacedName types.NamespacedName) {
 		By("Starting reconciliation process")
-		reconciler := &ClusterIPAMClaimReconciler{Client: k8sClient, Scheme: k8sClient.Scheme()}
+		reconciler := &ipam.ClusterIPAMClaimReconciler{Client: k8sClient, Scheme: k8sClient.Scheme()}
 		_, err := reconciler.Reconcile(ctx, reconcile.Request{NamespacedName: namespacedName})
 		Expect(err).NotTo(HaveOccurred())
