+++ test/functional/servicetemplate_controller_test.go
