+++ test/functional/template_controller_test.go
