+++ test/functional/templatechain_controller_test.go
